requirements.txt
astral==3.2
influxdb-client==1.36.1
pytz==2023.3
requests==2.30.0
PyYAML==6.0